requirements.txt
fastapi==0.121.3
mlflow==3.4.0
mlflow_skinny==3.4.0
mlflow_tracing==3.4.0
pandas==2.3.3
pydantic==2.12.4
python-dotenv==1.2.1
scikit_learn==1.7.2
uvicorn==0.38.0
xgboost==3.1.1
